Simply activate/deactivate conda
Instead of trying to manually add `conda` to and from the `PATH`, simply
use the `activate`/`deactivate` scripts to do this for us. Provide full
paths to the `activate` script to handle this change. Also run
`deactivate` at the end to make sure that one install of `conda` does
not mess with the install of the other `conda`. Finally recent versions
of `conda` don't need `root` specified and will simply default to `root`
if no environment is specified.