Activate the default `conda`'s `base` environment
Make sure to activate the default `conda`'s `base` environment as part
of the Docker entrypoint. This replaces kludgy hacks like were done
previously to trick `conda` into being active via environment variables.
Instead this actually activates `conda` making sure it is setup the way
it expects. This should work better with activation scripts that
packages might provide. Also should be ensure `conda` behaves how it is
suppose to in the full range of cases that it hadn't (in our case)
before.