Drop outdated comment
We plan to keep the Python 3 `conda` install as default. Though we may
drop the Python 2 `conda` install in the future. So this symlink may go,
but a lot of other stuff that is not commented will go with it. Further
will be adding changes here that are necessary for `conda` 4.4+'s
activation process and will be needed here permanently. For now drop the
comment to avoid confusion.