Export all environment variables `conda` sets
In order to ensure that all environment variables that `conda` sets are
exported, use the `set -a` option before calling `conda`. This tells
`bash` to `export` all environment variables that are assigned after the
`set -a` builtin is run. After `conda activate` has completed, `set +a`
is run to ensure that any later assignments are not automatically
exported. Thus when running this entrypoint script as part of a Docker
container, the user is dropped into an environment with `conda` already
activated.

ref: http://man7.org/linux/man-pages/man1/set.1p.html
ref: http://pubs.opengroup.org/onlinepubs/9699919799.2016edition/utilities/V3_chap02.html#set