Drop symlink of `conda.sh` into `/etc/profile.d/`
As we can now properly activate `conda` in the entrypoint script, there
is no need to rely on `/etc/profile.d/` scripts to play a roll in
activation of `conda`'s base environment. So drop the symlink of
`conda.sh` into `/etc/profile.d/`. This can just be run in the
entrypoint script instead.