Use new time server to bust cache
Appears that USNO's time server redirects to `https`, but their
certificates are invalid for some reason. Probably a consequence of
their recent move to `https`. Would expect them to fix this in the
future. For now, switch to a new time server to bust the cache. This
provides the output in JSON with a few formats as opposed to a simple
string, but is otherwise just text. Should work for our needs.