Share the Conda package cache between installs
The Conda packages between the Python 2 and Python 3 installs repeat
each other quite a bit (particularly for packages that are not dependent
on Python like `zlib`). As Conda is able to cutdown on package storage
overhead by using links (soft or hard) to the package cache directory,
it should be possible to cutdown significantly on image size by reusing
the package cache between the two Conda installs. Also should save on
build time in cases where the packages have already been solved for,
downloaded, and unpacked.

Given this, attempt to prep the Conda install locations for Python 2 and
Python 3 to ensure the package cache directory is shared between the two
installs. Also update the installation steps to ensure Miniconda is
comfortable installing into an already existent directory. These seem to
be the minimally invasive changes (as opposed to say moving files around
on disk). This seems to have the intended effect of sharing the package
cache directory between installs and cutting down on the overhead of
having two package caches.