Ensure that `python` comes from `conda-forge` only
Use `conda`'s new MatchSpec feature to pin `python` to the `conda-forge`
channel. As some of the installation steps were causing `python` to get
pulled from `defaults` instead (possibly due to its somewhat outdated
pinned dependencies), this is needed to ensure that `python` only comes
from `conda-forge`. Also this is a bit more effective at pinning
`python` version in the environment thanks to MatchSpec. As the
environment inherits this pinning, we can be sure that this constraint
will be upheld for all later installation steps within this Docker image
build or later images.