Remove the `defaults` channel
At this point, we can get basically everything except for a couple
packages from `conda-forge`. In fact, this has been the case for a long
time. However it has been convenient to build the image with merely the
conda-forge channel on top. Recently that has become problematic however
as things like `blas` or even `python` try to come from the `defaults`
channel for unclear reasons. To simplify things and ensure
reproducibility, simply remove the `defaults` channel. This way we can
explicitly install anything that must come from `defaults` and not worry
about other things getting pulled in accidentally.